requirements.txt
flask
numpy
pandas
plotly
flask-nav
jsons
pip
requests
paho-mqtt